splitted up importer scripts (#17)
splitted up importer scripts
